pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "sma-outfits"
version = "0.1.0"
description = "Alpaca-only recreation stack for SMA outfits"
readme = "README.md"
requires-python = "==3.14.3"
dependencies = [
  "duckdb>=1.1.3",
  "kaleido>=0.2.1",
  "numpy>=2.2.2",
  "pandas>=2.2.3",
  "plotly>=6.0.0",
  "pyarrow>=19.0.0",
  "pydantic>=2.10.0",
  "PyYAML>=6.0.2",
  "requests>=2.32.3",
  "typer>=0.12.5",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.4",
]

[project.scripts]
sma-outfits = "sma_outfits.cli:app"

[tool.pytest.ini_options]
addopts = "-q"
testpaths = ["tests"]
